refactor(admin): 修改 UploadService接口中的方法参数名称
- 将 testBaseInfoAccess 方法中的参数名从 uploadInfo 改为 uploadBaseInfo
- 这个改动提高了代码的可读性和一致性，使其更符合实际的业务含义

+++ src/main/java/com/mysiteforme/admin/entity/PermissionButton.java
@@ -42,7 +42,7 @@ public class PermissionButton extends DataEntity {
      * 排序
      */
 	@TableField("sort")
-	private Short sort;
+	private Integer sort;
